src/requirements.txt
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dash==2.17.0
pandas
